[NS-72] Clean up `LocalVeoExtension`
- Migrate `LocalVeoExtension` over to `io.sparkcyclone.plugin`

- Migrate `CombinedCompilationColumnarRule`, `VeColumnarRule`, and
`VeHintResolutionRule` over to `io.sparkcyclone.spark.rules`
