+++ src/test/scala/io/sparkcyclone/tpc/TPCHVESqlSpec.scala
@@ -20,9 +20,8 @@
 package io.sparkcyclone.tpc
 
 import io.sparkcyclone.plugin.{AuroraSqlPlugin, SparkCycloneExecutorPlugin}
-import io.sparkcyclone.spark.LocalVeoExtension.compilerRule
 import io.sparkcyclone.plugin.SparkCycloneExecutorPlugin.CloseAutomatically
-import io.sparkcyclone.spark.planning.{VERewriteStrategy, VeColumnarRule, VeRewriteStrategyOptions}
+import io.sparkcyclone.spark.planning.{VERewriteStrategy, VeRewriteStrategyOptions}
 import io.sparkcyclone.vectorengine.VeProcess
 import java.io.File
 import org.apache.spark.sql.SparkSession
